Build cleanup (#698)
* check if loading apps is still needed

* Revert "check if loading apps is still needed"

This reverts commit d03cfe562043805d55934bd770940a44cc27eb43.

* rename function to match what it actually do

* extract diagnostic related code

* extract common mix.exs code

* fix tests
