+++ apps/elixir_ls_utils/lib/mixfile_helpers.ex
@@ -1,0 +1,5 @@
+defmodule ElixirLS.Utils.MixfileHelpers do
+  def mix_exs do
+    System.get_env("MIX_EXS") || "mix.exs"
+  end
+end
