+++ apps/elixir_ls_utils/test/support/mix_test.case.ex
@@ -1,6 +1,7 @@
 defmodule ElixirLS.Utils.MixTest.Case do
   # This module is based heavily on MixTest.Case in Elixir's tests
   use ExUnit.CaseTemplate
+  alias ElixirLS.Utils.MixfileHelpers
 
   using do
     quote do
@@ -103,7 +104,7 @@ defmodule ElixirLS.Utils.MixTest.Case do
     clear_mix_cache()
 
     # Attempt to purge mixfiles for dependencies to avoid module redefinition warnings
-    mix_exs = System.get_env("MIX_EXS") || "mix.exs"
+    mix_exs = MixfileHelpers.mix_exs()
 
     for {mod, :in_memory} <- :code.all_loaded(),
         source = mod.module_info[:compile][:source],
