+++ apps/language_server/test/providers/code_lens/test_test.exs
@@ -7,7 +7,7 @@ defmodule ElixirLS.LanguageServer.Providers.CodeLens.TestTest do
   @project_dir "/project"
 
   setup context do
-    ElixirLS.LanguageServer.Build.load_all_modules()
+    ElixirLS.LanguageServer.Build.load_all_mix_applications()
 
     unless context[:skip_server] do
       server = ElixirLS.LanguageServer.Test.ServerTestHelpers.start_server()
